Fix some gtk changes.
- Use new gtr_* functions and changed variable names.
- Move setting boilerplate to correct place.
- Fix signed/unsigned comparison.

+++ gtk/tr-core.h
@@ -84,9 +84,9 @@ void           tr_core_clear( TrCore * self );
 
 tr_session *   tr_core_session( TrCore * self );
 
-int            tr_core_get_inactive_queued_torrent_count( TrCore * self );
+size_t         tr_core_get_inactive_queued_torrent_count( TrCore * self );
 
-int            tr_core_get_inactive_torrent_count( TrCore * self );
+size_t         tr_core_get_inactive_torrent_count( TrCore * self );
 
 size_t         tr_core_get_active_torrent_count( TrCore * self );
